[EGSVC-1842]: Vehicle Maintenance missing validations.

+++ swm/swm-services/src/main/java/org/egov/swm/persistence/entity/VehicleMaintenanceEntity.java
@@ -29,7 +29,7 @@ public class VehicleMaintenanceEntity {
 
     private String downtimeforMaintenanceUom = null;
 
-    private Long downtimeforMaintenance = null;
+    private Double downtimeforMaintenance = null;
 
     private String createdBy = null;
